feature: Add npm based webapps
Add npm based webapps, which can be installed with npm install and show up
under /apps, just like bower installed webapps have shown. SK webapp modules
are marked with signalk-webapp keyword in package.json. Bower and npm based
webapps can coexist. This change supports also npm modules that provide
both a webapp interface (their entire contents are available under the
server's html root) and a plugin interface.
