+++ lib/index.js
@@ -53,7 +53,30 @@ function Server(opts) {
     res.sendFile(__dirname + '/index.html')
   })
 
+  app.get('/apps', (req, res, next) => {
+    var html = fs.readFileSync(__dirname + '/appindex.html', {encoding: 'utf8'})
+    var insertionIndex = html.indexOf('</div>');
+    var sliceToInsertion = html.slice(0, insertionIndex);
+    var sliceToEnd = html.slice(insertionIndex);
+
+    var result = sliceToInsertion;
+    result += '<ul class="list-group">';
+    result += app.webapps.reduce(function(result, componentInfo) {
+      result += '<li class="list-group-item">';
+      result += '<b><a href="' + componentInfo.name + '">' + componentInfo.name + '</a></b> ';
+      result += componentInfo.description;
+      result += '</li>\n';
+      return result;
+    }, '');
+    result += '</ul>';
+    result += sliceToEnd;
+    res.send(result);
+  })
+
+
   app.signalk = new FullSignalK(app.selfId, app.selfType, app.config.defaults);
+
+  app.webapps = []
 
   app.streambundle = new StreamBundle(app.selfId);
   app.signalk.on('delta', app.streambundle.pushDelta.bind(app.streambundle))
