+++ lib/interfaces/bower.js
@@ -24,30 +24,10 @@ module.exports = function(app) {
     start: function() {
       app.use('/bower_components', express.static(__dirname + '/../../bower_components'));
       app.use('/', express.static(__dirname + '/../../bower_components'));
-      app.get('/apps', bowerIndex);
+      app.webapps = app.webapps.concat(getBowerComponents())
     }
   }
 };
-
-function bowerIndex(req, res, next) {
-  var html = require('fs').readFileSync(__dirname + '/bowerindex.html', {encoding: 'utf8'})
-  var insertionIndex = html.indexOf('</div>');
-  var sliceToInsertion = html.slice(0, insertionIndex);
-  var sliceToEnd = html.slice(insertionIndex);
-
-  var result = sliceToInsertion;
-  result += '<ul class="list-group">';
-  result += getBowerComponents().reduce(function(result, componentInfo) {
-    result += '<li class="list-group-item">';
-    result += '<b><a href="' + componentInfo.name + '">' + componentInfo.name + '</a></b> ';
-    result += componentInfo.description;
-    result += '</li>\n';
-    return result;
-  }, '');
-  result += '</ul>';
-  result += sliceToEnd;
-  res.send(result);
-}
 
 function getBowerComponents() {
   var bowerBaseDir = __dirname + '/../../bower_components/';
